Allan
Random Commit #725

+++ Unity Base Project/Assets/Scripts/CSharp/Properties/HealthProperties.cs
@@ -6,7 +6,7 @@ public class HealthProperties
 {
 
     [SerializeField]
-    public int Health { get; private set; }
+    public float Health { get; private set; }
 
     private Transform enemyRef;
 
@@ -15,13 +15,13 @@ public class HealthProperties
         
     }
 
-    public void Set(int _hp, Transform _ref)
+    public void Set(float _hp, Transform _ref)
     {
         Health = _hp;
         enemyRef = _ref;
     }
 
-    public void Damage(int _dmg)
+    public void Damage(float _dmg)
     {
         Health -= _dmg;
         if (Health <= 0)
